+++ app/src/main/java/com/elhady/movies/data/local/AppConfigurationImp.kt
@@ -1,4 +1,4 @@
-package com.elhady.movies.data
+package com.elhady.movies.data.local
 
 import javax.inject.Inject
