+++ app/src/main/java/com/elhady/movies/data/local/database/MovieDatabase.kt
@@ -1,10 +1,11 @@
-package com.elhady.movies.data.database
+package com.elhady.movies.data.local.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.elhady.movies.data.database.daos.MovieDao
-import com.elhady.movies.data.database.entity.PopularMovieEntity
+import com.elhady.movies.data.local.database.daos.MovieDao
+import com.elhady.movies.data.local.database.entity.PopularMovieEntity
+import com.elhady.movies.data.local.Converters
 
 @TypeConverters(Converters::class)
 @Database(entities = [PopularMovieEntity::class], version = 1)
