+++ app/src/main/java/com/elhady/movies/data/local/database/daos/MovieDao.kt
@@ -1,10 +1,10 @@
-package com.elhady.movies.data.database.daos
+package com.elhady.movies.data.local.database.daos
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.elhady.movies.data.database.entity.PopularMovieEntity
+import com.elhady.movies.data.local.database.entity.PopularMovieEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
