+++ app/src/main/java/com/elhady/movies/data/repository/MovieRepository.kt
@@ -1,13 +1,11 @@
 package com.elhady.movies.data.repository
 
-import com.elhady.movies.data.database.entity.PopularMovieEntity
+import com.elhady.movies.data.local.database.entity.PopularMovieEntity
 import com.elhady.movies.data.remote.State
 import com.elhady.movies.data.remote.response.BaseResponse
 import com.elhady.movies.data.remote.response.MovieDto
 import com.elhady.movies.data.remote.response.PersonDto
 import com.elhady.movies.data.remote.response.genre.GenreDto
-import com.elhady.movies.data.remote.response.genre.GenreResponse
-import com.elhady.movies.domain.models.PopularMovie
 import kotlinx.coroutines.flow.Flow
 
 interface MovieRepository {
