+++ app/src/main/java/com/elhady/movies/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.elhady.movies.di
 
-import com.elhady.movies.data.AppConfiguration
+import com.elhady.movies.data.local.AppConfiguration
 import com.elhady.movies.data.DataClassParser
-import com.elhady.movies.data.database.daos.MovieDao
+import com.elhady.movies.data.local.database.daos.MovieDao
 import com.elhady.movies.data.repository.MovieRepository
 import com.elhady.movies.data.repository.MovieRepositoryImp
 import com.elhady.movies.data.remote.service.MovieService
