+++ app/src/main/java/com/elhady/movies/domain/mappers/MovieMapper.kt
@@ -1,6 +1,6 @@
 package com.elhady.movies.domain.mappers
 
-import com.elhady.movies.data.database.entity.PopularMovieEntity
+import com.elhady.movies.data.local.database.entity.PopularMovieEntity
 import com.elhady.movies.data.remote.response.MovieDto
 import com.elhady.movies.data.remote.response.genre.GenreDto
 import com.elhady.movies.utilities.Constants
